chore: adjust the student relation to site id so that it does not delete the student record
found in testing that it was set up to delete the student record when a site was deleted, this
should allow for sites to be deleted without removing our students. If a site is deleted now it will
be possible to re-fill in the site they're associated with if the reference point changes.